LPS-79588 Avoid NPE

+++ modules/apps/message-boards/message-boards-service/src/main/java/com/liferay/message/boards/service/impl/MBMessageLocalServiceImpl.java
@@ -1709,7 +1709,9 @@ public class MBMessageLocalServiceImpl extends MBMessageLocalServiceBaseImpl {
 					String fileEntryId = String.valueOf(
 						fileEntry.getFileEntryId());
 
-					if (!existingFiles.contains(fileEntryId)) {
+					if ((existingFiles != null) &&
+						!existingFiles.contains(fileEntryId)) {
+
 						if (!TrashUtil.isTrashEnabled(message.getGroupId())) {
 							deleteMessageAttachment(
 								messageId, fileEntry.getTitle());
